"Deploy" JHDF5 locally
The JHDF5 library is only a wrapper around a JNI library, so it is hardly
"HDF5 for Java" since they cannot hope to begin to support as many
platforms as Java does. So we downloaded the library from

	https://wiki-bsse.ethz.ch/pages/viewpage.action?pageId=26609113

and used the trick mentioned in

	http://blog.dub.podval.org/2010/01/maven-in-project-repository.html

to "deploy" the .jar file containing native libraries for at least almost
all platforms supported by Fiji (PowerPC MacOSX is missing), using this
command line:

	mvn install:install-file \
		-Dfile=$HOME/cisd-jhdf5-batteries_included_lin_win_mac.jar \
		-DgroupId=cisd \
		-DartifactId=jhdf5 \
		-Dversion=12.02.3 \
		-Dpackaging=jar \
		-DgeneratePom=true \
		-DlocalRepositoryPath=$(pwd)/jhdf5

We will need to come up with a portable solution rather than JHDF5,
though.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>