Replace fetcher BlockingDeque by (blocking) set of queues. The Queues have different priorities, blocks with highest priority (lowest res) are loaded first.
Queues are cleared for every new frame (working through screenscales
counts as a single frame) such that the blocks required for the current
frame are loaded immediately. The plan is to move the cleared-out requests
to a prefetch deque of limited size.

For queue-clearing, SpimViewer needs to know the Hdf5GlobalCache which is
not so nice. Maybe we can find a better solution later.