Use imglib2-cache to make per-image VolatileLoadingCaches
There is one VolatileLoadingCache per image with a loader that simply takes
a Long cell index. This wraps a VolatileCache (always use the same loader).
The VolatileCache is backed by a global Cache to which it adapts by a
KeyBimap that translate between Long keys and Key objects that also have
setup/timepoint/resolution information.

Fix resulting ImageLoader errors.