Avoid exception when canvas size is zero
I ran into a case where BehaviourTransformEventHandler3D.setCanvasSize
gets called with (0, 0) size. When this occurs, we should perform a new
transformation because it will result in a "Matrix is singular"
exception. In practice, this call happens at an intermediate layout
step, after which the canvas will be resized again to non-zero size.