Fix setting of CacheHints prefetch and render
With the DefaultMipMapOrdering these are both null, but we still want to
do something reasonable. Before this was just controlled by setting the
I/O time budget. But this breaks if the default LoadingStrategy of the
image is BLOCKING (in this case prefetching is anyway not very useful,
but still, this new solution is less brittle).