Add ViewerOption.shareKeyPressedEvents(...) for multi-window ui-behaviour
Background: Multiple ui-behaviour windows can share one KeyPressManager.
keyPressed() events are forwarded to and distributed by the manager. This
is intended to make keyboard click/drag behaviours work like mouse
click/drag: When a behaviour is initiated with a key press, the window
under the mouse receives focus and the behaviour is handled there.