Use imglib2-cache-beta-12 and fix compile errors.
In particular, BlockingFetchQueues now needs to know the number of
consumers (fetcher-threads). To realize cache invalidate, it must be
possible to pause the fetcher-threads, which means driving the queue into
a state where all consumers are held in take(). To do that,
BlockingFetchQueues.pause() blocks until all (of the pre-defined number
of) consumer threads have arrived in take().