MultiResolutionRenderer: Improve heuristic for screen scale selection
This fixes the problem described in paintera source code:

1) With vastly different screen scale values such as [1.0, 0.1], it will switch between the two scales every frame.
At screen scale 0.1 rendering is fast, and after a frame is rendered, it switches maxScreenScaleIndex to 0.
However, at screen scale 1.0 rendering is slower than targetRenderNanos, so when the next frame is rendered, it switches maxScreenScaleIndex back to 1.
This causes annoying delays because when maxScreenScaleIndex is switched to 0, renderingMayBeCancelled is in turn set to false, and the algorithm has to wait
until the current high-res frame is fully rendered.

2) When the user starts painting, it re-renders only the affected interval which is usually very small compared to the size of the screen.
Usually rendering is very fast in this case, and maxScreenScaleIndex is quickly changed to 0.
When the user finishes painting and starts navigating again, there may be a delay in rendering the first few frames because
it starts from the highest available resolution and then gradually decreases the resolution until the rendertime is within the targetRenderNanos threshold.