RealARGBColorConverter: fix memory leak when calling ClassCopyProvider
This line causes BDV to keep a reference to the first (or more) images
shown. ClassCopyProvider keeps the keys used for creating class copies.
In the key was a ImgLib2 pixel. So class copy kept a reference to the
pixel. The pixel kept a reference to it's image, which than never gets
garbage collected.