Add more datatypes to Hdf5ImageLoader
For this to work, the datatype must be added as an attribute to each setup
group. E.g., the info (pyramid resolutions, etc) for the first setup is
under group "/s00" in the h5 file. If "/s00" has a "dataType" String
attribute with value "uint8", that means that the datasets for this setup
are UnsignedByteType. The values for the "dataType" attribute are the same
as for N5. See DataType.toString(). If no "dataType" attribute is present,
it is assumed that the datasets for this setup are legacy
UnsignedShortType-stored-as-INT16.

Supported datatypes are INT8, UINT8, INT16, UINT16, INT32, UINT32, INT64,
UINT64, FLOAT32, FLOAT64.