Add JsonUtils in particular @JsonIo annotation
Typed<T> instances of any T with a @JsonIo-annotated adapter are
serialized as json objects with a "type" attribute with the value of the
JsonIo.type() annotation, and an "obj" attribute, which is serialized
using the annotated adapter. For deserialization, the correct adapter is
looked up via the "type" attribute.